requirements.txt
flask==2.2.3
flask-cors==3.0.10
pillow==9.4.0
werkzeug==2.2.3
gunicorn==20.1.0
requests==2.28.2
pytesseract==0.3.10  # Mantén esto si vas a intentar instalar Tesseract
